Wrapping up the project so that I can focus on job interviews and their concomitant prep. Note that my blinking LED code and their assays with the NUCLEO-F767ZI PWM outputs for triangular and sinusoidal waveforms have all been tested and proven working.
Only thing remaining to be debugged and tested are the third-party drivers for the DHT11 sensor and the LCD16x2 display. But since these are third-party libraries and they were originally developed for Mbed 2 hence obsolete, I can hardly be blamed for them not currently working.

A future project for myself when time permits is to write my own drivers for these sensors. That way, I am guaranteed knowing that their driver code is indeed working. So many projects, not enough time!